try fully sequential
